Added Materials + Bugfixes
Added the .materals folder, though there's nothing new in it. To be used
in future updates. Fixed various bugs: "read" command now reads instead
of examining. The default printouts when the player succeeds at an
action are in the correct tense and refer to the player character
instead of "you". This also fixed the error where reading the letter
resulted in blank space instead of the contents of the letter.